refactor: 파일 이름 규칙 변경, utils 폴더 분할
- 가독성 개선을 위해 파일명 camelCase -> kebab-case 컨벤션 적용
- utils 폴더에 모인 파일들을 아래 폴더들로 분할
  -  handlers 폴더: 단순 기능 함수
  -  contollers 폴더: 이벤트 등록 등 처리하는 파일
