+++ src/handlers/focus-handler.js
