+++ src/handlers/mobile-recognizer.js
