+++ src/handlers/section-scroller.js
